mark Diff::rollback as unimplemented
I have marked the rollback function in the diff module as unimplemented.
Trying to write it and rewrite it has made me realize that my
representation of an edit is pretty janky and flawed.

Since I'm going to be refactoring essentially this entire module if I
refactor the Edit struct, it would be worthless to try and write the
rollback function. It will also be much simpler to write with the
refactored Edit struct.