Change the representation of an edit
This representation of an edit should be a lot more complete, and also
make a lot of code a lot more concise and less error-prone.

The edit now references both the original and modified file's line
numbers and content. This allows makes the line numbers in an insert and
delete edit a lot more clear, and allows for a replacement operation,
a lack of which was a major issue in the rollback function

I've also tried to make illegal types unrepresentable, so the HalfEdit
struct does not have a line end field but is instead based on the length
of the content field, which is now a vec of Strings instead of just a
single String.

There are some invalud types that are still representable, but if I had
implemented it so that it wasn't that way, later code would be pretty
janky, so I made the tradeoff of cleaner code for a possible invalid
representation of a private struct, which I think is fine.