Re-add commit write step
Actually, having a write function is nice because we can create a commit
for the purposes of diffing, ex. the state of the repository vs the last
commit to find changes.

Some aspects like passing the path of the original file back into
Content::write are not ideal, and making a new struct for the commit
tree and 'promoting' it into an actual commit while writing it to disk
would probably be better, but it also requires a lot of rewriting...

maybe later.